Write seitan v1 by default, remove log
