+++ go/protocol/keybase1/teams.go
@@ -2157,7 +2157,7 @@ type TeamCreateSeitanTokenArg struct {
 	Label     SeitanKeyLabel `codec:"label" json:"label"`
 }
 
-type TeamCreateSeitanTokenV1Arg struct {
+type TeamCreateSeitanTokenV2Arg struct {
 	SessionID int            `codec:"sessionID" json:"sessionID"`
 	Name      string         `codec:"name" json:"name"`
 	Role      TeamRole       `codec:"role" json:"role"`
@@ -2263,8 +2263,8 @@ type TeamsInterface interface {
 	TeamTree(context.Context, TeamTreeArg) (TeamTreeResult, error)
 	TeamDelete(context.Context, TeamDeleteArg) error
 	TeamSetSettings(context.Context, TeamSetSettingsArg) error
-	TeamCreateSeitanToken(context.Context, TeamCreateSeitanTokenArg) (SeitanIKeyV2, error)
-	TeamCreateSeitanTokenV1(context.Context, TeamCreateSeitanTokenV1Arg) (SeitanIKey, error)
+	TeamCreateSeitanToken(context.Context, TeamCreateSeitanTokenArg) (SeitanIKey, error)
+	TeamCreateSeitanTokenV2(context.Context, TeamCreateSeitanTokenV2Arg) (SeitanIKeyV2, error)
 	TeamAddEmailsBulk(context.Context, TeamAddEmailsBulkArg) (BulkRes, error)
 	LookupImplicitTeam(context.Context, LookupImplicitTeamArg) (LookupImplicitTeamRes, error)
 	LookupOrCreateImplicitTeam(context.Context, LookupOrCreateImplicitTeamArg) (LookupImplicitTeamRes, error)
@@ -2673,18 +2673,18 @@ func TeamsProtocol(i TeamsInterface) rpc.Protocol {
 				},
 				MethodType: rpc.MethodCall,
 			},
-			"teamCreateSeitanTokenV1": {
+			"teamCreateSeitanTokenV2": {
 				MakeArg: func() interface{} {
-					ret := make([]TeamCreateSeitanTokenV1Arg, 1)
+					ret := make([]TeamCreateSeitanTokenV2Arg, 1)
 					return &ret
 				},
 				Handler: func(ctx context.Context, args interface{}) (ret interface{}, err error) {
-					typedArgs, ok := args.(*[]TeamCreateSeitanTokenV1Arg)
+					typedArgs, ok := args.(*[]TeamCreateSeitanTokenV2Arg)
 					if !ok {
-						err = rpc.NewTypeError((*[]TeamCreateSeitanTokenV1Arg)(nil), args)
+						err = rpc.NewTypeError((*[]TeamCreateSeitanTokenV2Arg)(nil), args)
 						return
 					}
-					ret, err = i.TeamCreateSeitanTokenV1(ctx, (*typedArgs)[0])
+					ret, err = i.TeamCreateSeitanTokenV2(ctx, (*typedArgs)[0])
 					return
 				},
 				MethodType: rpc.MethodCall,
@@ -3053,13 +3053,13 @@ func (c TeamsClient) TeamSetSettings(ctx context.Context, __arg TeamSetSettingsA
 	return
 }
 
-func (c TeamsClient) TeamCreateSeitanToken(ctx context.Context, __arg TeamCreateSeitanTokenArg) (res SeitanIKeyV2, err error) {
+func (c TeamsClient) TeamCreateSeitanToken(ctx context.Context, __arg TeamCreateSeitanTokenArg) (res SeitanIKey, err error) {
 	err = c.Cli.Call(ctx, "keybase.1.teams.teamCreateSeitanToken", []interface{}{__arg}, &res)
 	return
 }
 
-func (c TeamsClient) TeamCreateSeitanTokenV1(ctx context.Context, __arg TeamCreateSeitanTokenV1Arg) (res SeitanIKey, err error) {
-	err = c.Cli.Call(ctx, "keybase.1.teams.teamCreateSeitanTokenV1", []interface{}{__arg}, &res)
+func (c TeamsClient) TeamCreateSeitanTokenV2(ctx context.Context, __arg TeamCreateSeitanTokenV2Arg) (res SeitanIKeyV2, err error) {
+	err = c.Cli.Call(ctx, "keybase.1.teams.teamCreateSeitanTokenV2", []interface{}{__arg}, &res)
 	return
 }
