+++ go/service/teams.go
@@ -369,7 +369,7 @@ func (h *TeamsHandler) LoadTeamPlusApplicationKeys(ctx context.Context, arg keyb
 	return teams.LoadTeamPlusApplicationKeys(ctx, h.G().ExternalG(), arg.Id, arg.Application, arg.Refreshers)
 }
 
-func (h *TeamsHandler) TeamCreateSeitanToken(ctx context.Context, arg keybase1.TeamCreateSeitanTokenArg) (token keybase1.SeitanIKeyV2, err error) {
+func (h *TeamsHandler) TeamCreateSeitanToken(ctx context.Context, arg keybase1.TeamCreateSeitanTokenArg) (token keybase1.SeitanIKey, err error) {
 	ctx = libkb.WithLogTag(ctx, "TM")
 	if err := h.assertLoggedIn(ctx); err != nil {
 		return "", err
@@ -377,13 +377,12 @@ func (h *TeamsHandler) TeamCreateSeitanToken(ctx context.Context, arg keybase1.T
 	return teams.CreateSeitanToken(ctx, h.G().ExternalG(), arg.Name, arg.Role, arg.Label)
 }
 
-func (h *TeamsHandler) TeamCreateSeitanTokenV1(ctx context.Context, arg keybase1.TeamCreateSeitanTokenV1Arg) (token keybase1.SeitanIKey, err error) {
-	// NOTE: Only used for integration tests until the deprecation of seitanv1
+func (h *TeamsHandler) TeamCreateSeitanTokenV2(ctx context.Context, arg keybase1.TeamCreateSeitanTokenV2Arg) (token keybase1.SeitanIKeyV2, err error) {
 	ctx = libkb.WithLogTag(ctx, "TM")
 	if err := h.assertLoggedIn(ctx); err != nil {
 		return "", err
 	}
-	return teams.CreateSeitanTokenV1(ctx, h.G().ExternalG(), arg.Name, arg.Role, arg.Label)
+	return teams.CreateSeitanTokenV2(ctx, h.G().ExternalG(), arg.Name, arg.Role, arg.Label)
 }
 
 func (h *TeamsHandler) GetTeamRootID(ctx context.Context, id keybase1.TeamID) (keybase1.TeamID, error) {
