+++ go/systests/team_seitan_invite_test.go
@@ -1,105 +1,16 @@
 package systests
 
 import (
-	"fmt"
 	"strconv"
 	"testing"
 	"time"
 
-	"golang.org/x/net/context"
-
-	libkb "github.com/keybase/client/go/libkb"
+	"github.com/keybase/client/go/libkb"
 	keybase1 "github.com/keybase/client/go/protocol/keybase1"
 	"github.com/keybase/client/go/teams"
 	"github.com/stretchr/testify/require"
+	context "golang.org/x/net/context"
 )
-
-func TestTeamInviteSeitanHappy(t *testing.T) {
-	testTeamInviteSeitanHappy(t, false /* implicitAdmin */, true /* seitanV1 */)
-	testTeamInviteSeitanHappy(t, false /* implicitAdmin */, false /* seitanV1 */)
-}
-
-func TestTeamInviteSeitanHappyImplicitAdmin(t *testing.T) {
-	testTeamInviteSeitanHappy(t, true /* implicitAdmin */, true /* seitanV1 */)
-	testTeamInviteSeitanHappy(t, true /* implicitAdmin */, false /* seitanV1 */)
-}
-
-func testTeamInviteSeitanHappy(t *testing.T, implicitAdmin bool, seitanV1 bool) {
-	tt := newTeamTester(t)
-	defer tt.cleanup()
-
-	own := tt.addUser("own")
-	roo := tt.addUser("roo")
-
-	teamIDParent, teamNameParent := own.createTeam2()
-	teamID := teamIDParent
-	teamName := teamNameParent
-	t.Logf("Created team %v %v", teamIDParent, teamNameParent)
-	if implicitAdmin {
-		subteamID, err := teams.CreateSubteam(context.TODO(), tt.users[0].tc.G, "sub1", teamNameParent)
-		require.NoError(t, err)
-		teamID = *subteamID
-		subteamName, err := teamNameParent.Append("sub1")
-		require.NoError(t, err)
-		teamName = subteamName
-		t.Logf("Created subteam %v %v", teamID, teamName)
-	}
-
-	label := keybase1.NewSeitanKeyLabelWithSms(keybase1.SeitanKeyLabelSms{
-		F: "bugs",
-		N: "0000",
-	})
-	var token string
-	if seitanV1 {
-		ikey, err := own.teamsClient.TeamCreateSeitanTokenV1(context.TODO(), keybase1.TeamCreateSeitanTokenV1Arg{
-			Name:  teamName.String(),
-			Role:  keybase1.TeamRole_WRITER,
-			Label: label,
-		})
-		token = string(ikey)
-		require.NoError(t, err)
-	} else {
-		ikey, err := own.teamsClient.TeamCreateSeitanToken(context.TODO(), keybase1.TeamCreateSeitanTokenArg{
-			Name:  teamName.String(),
-			Role:  keybase1.TeamRole_WRITER,
-			Label: label,
-		})
-		token = string(ikey)
-		require.NoError(t, err)
-	}
-
-	t.Logf("Created token %q", token)
-
-	details := own.teamGetDetails(teamName.String())
-	require.Len(t, details.AnnotatedActiveInvites, 1)
-	for _, invite := range details.AnnotatedActiveInvites {
-		require.Equal(t, keybase1.TeamRole_WRITER, invite.Role)
-		tic, err := invite.Type.C()
-		require.NoError(t, err)
-		require.Equal(t, keybase1.TeamInviteCategory_SEITAN, tic)
-		require.Equal(t, keybase1.TeamInviteName("bugs (0000)"), invite.Name)
-	}
-
-	err := roo.teamsClient.TeamAcceptInvite(context.TODO(), keybase1.TeamAcceptInviteArg{
-		Token: token,
-	})
-	require.NoError(t, err)
-
-	t.Logf("User used token, waiting for rekeyd")
-
-	own.kickTeamRekeyd()
-	own.waitForTeamChangedGregor(teamID, keybase1.Seqno(3))
-
-	t0, err := teams.GetTeamByNameForTest(context.TODO(), own.tc.G, teamName.String(), false /* public */, true /* needAdmin */)
-	require.NoError(t, err)
-
-	role, err := t0.MemberRole(context.TODO(), teams.NewUserVersion(roo.uid, 1))
-	require.NoError(t, err)
-	require.Equal(t, role, keybase1.TeamRole_WRITER)
-
-	details = own.teamGetDetails(teamName.String())
-	require.Len(t, details.AnnotatedActiveInvites, 0)
-}
 
 func TestTeamInviteSeitanFailures(t *testing.T) {
 	tt := newTeamTester(t)
@@ -120,9 +31,9 @@ func TestTeamInviteSeitanFailures(t *testing.T) {
 
 	t.Logf("Created token %q", token)
 
-	// Generate invitation id, but make Signature with different IKey.
+	// Generate invitation id, but make AKey with different IKey.
 	// Simulate "replay attack" or similar.
-	ikey, err := teams.GenerateIKeyV2FromString(string(token))
+	ikey, err := teams.GenerateIKeyFromString(string(token))
 	require.NoError(t, err)
 	sikey, err := ikey.GenerateSIKey()
 	require.NoError(t, err)
@@ -131,22 +42,23 @@ func TestTeamInviteSeitanFailures(t *testing.T) {
 	inviteID, err := keybase1.TeamInviteIDFromString(string(inviteIDx))
 	require.NoError(t, err)
 
-	ikey2, err := teams.GenerateIKeyV2()
+	ikey2, err := teams.GenerateIKey()
 	require.NoError(t, err)
 	sikey2, err := ikey2.GenerateSIKey()
 	require.NoError(t, err)
-	now := keybase1.ToTime(time.Now())
-	_, maliciousPayload, err := sikey2.GenerateSignature(roo.uid, roo.userVersion().EldestSeqno, teams.SCTeamInviteID(inviteID), now)
+	unixNow := time.Now().Unix()
+	_, maliciousPayload, err := sikey2.GenerateAcceptanceKey(roo.uid, roo.userVersion().EldestSeqno, unixNow)
 	require.NoError(t, err)
 
-	arg := libkb.NewAPIArgWithNetContext(context.Background(), "team/seitan_v2")
+	arg := libkb.NewAPIArgWithNetContext(context.Background(), "team/seitan")
 	arg.Args = libkb.NewHTTPArgs()
 	arg.SessionType = libkb.APISessionTypeREQUIRED
-	arg.Args.Add("sig", libkb.S{Val: maliciousPayload})
-	arg.Args.Add("now", libkb.S{Val: strconv.FormatInt(int64(now), 10)})
+	arg.Args.Add("akey", libkb.S{Val: maliciousPayload})
+	arg.Args.Add("now", libkb.S{Val: strconv.FormatInt(unixNow, 10)})
 	arg.Args.Add("invite_id", libkb.S{Val: string(inviteID)})
 	_, err = roo.tc.G.API.Post(arg)
 	require.NoError(t, err)
+
 	t.Logf("handle synthesized rekeyd command")
 	msg := keybase1.TeamSeitanMsg{
 		TeamID: teamID,
@@ -156,7 +68,7 @@ func TestTeamInviteSeitanFailures(t *testing.T) {
 			EldestSeqno: roo.userVersion().EldestSeqno,
 			Akey:        keybase1.SeitanAKey(maliciousPayload),
 			Role:        keybase1.TeamRole_WRITER,
-			UnixCTime:   int64(now),
+			UnixCTime:   unixNow,
 		}},
 	}
 	err = teams.HandleTeamSeitan(context.Background(), own.tc.G, msg)
@@ -172,84 +84,4 @@ func TestTeamInviteSeitanFailures(t *testing.T) {
 	role, err := t0.MemberRole(context.Background(), roo.userVersion())
 	require.NoError(t, err)
 	require.Equal(t, keybase1.TeamRole_NONE, role, "user role")
-}
-
-func testTeamCreateSeitanAndCancel(t *testing.T, seitanV1 bool) {
-	tt := newTeamTester(t)
-	defer tt.cleanup()
-
-	own := tt.addUser("own")
-
-	_, teamName := own.createTeam2()
-
-	t.Logf("Created team %q", teamName.String())
-
-	var labelSms keybase1.SeitanKeyLabelSms
-	labelSms.F = "Patricia S. Goldman-Rakic"
-	labelSms.N = "+481II222333"
-	label := keybase1.NewSeitanKeyLabelWithSms(labelSms)
-
-	var err error
-	if seitanV1 {
-		_, err = own.teamsClient.TeamCreateSeitanTokenV1(context.TODO(), keybase1.TeamCreateSeitanTokenV1Arg{
-			Name:  teamName.String(),
-			Role:  keybase1.TeamRole_WRITER,
-			Label: label,
-		})
-	} else {
-		_, err = own.teamsClient.TeamCreateSeitanToken(context.TODO(), keybase1.TeamCreateSeitanTokenArg{
-			Name:  teamName.String(),
-			Role:  keybase1.TeamRole_WRITER,
-			Label: label,
-		})
-	}
-	require.NoError(t, err)
-
-	t.Logf("Created Seitan token")
-
-	details, err := own.teamsClient.TeamGet(context.TODO(), keybase1.TeamGetArg{
-		Name:        teamName.String(),
-		ForceRepoll: true,
-	})
-	require.NoError(t, err)
-
-	var inviteID keybase1.TeamInviteID
-
-	require.Equal(t, 1, len(details.AnnotatedActiveInvites))
-	for key, invite := range details.AnnotatedActiveInvites {
-		require.Equal(t, keybase1.TeamRole_WRITER, invite.Role)
-		require.EqualValues(t, fmt.Sprintf("%s (%s)", labelSms.F, labelSms.N), invite.Name)
-
-		category, err := invite.Type.C()
-		require.NoError(t, err)
-		require.Equal(t, keybase1.TeamInviteCategory_SEITAN, category)
-
-		// Test rest of the params, unrelated to Seitan.
-		require.Equal(t, key, invite.Id)
-		require.Equal(t, keybase1.UserVersion{}, invite.Uv)
-		require.Equal(t, keybase1.UserVersion{Uid: own.uid, EldestSeqno: 1}, invite.Inviter)
-		require.Equal(t, own.username, invite.InviterUsername)
-		require.Equal(t, teamName.String(), invite.TeamName)
-
-		inviteID = invite.Id
-	}
-
-	t.Logf("Checked that invite was added correctly, removing invite by id")
-
-	err = own.teamsClient.TeamRemoveMember(context.TODO(), keybase1.TeamRemoveMemberArg{
-		Name:     teamName.String(),
-		InviteID: inviteID,
-	})
-	require.NoError(t, err)
-
-	t.Logf("Removed, checking if there are no active invites")
-
-	t0, err := teams.GetTeamByNameForTest(context.TODO(), own.tc.G, teamName.String(), false /* public */, true /* needAdmin */)
-	require.NoError(t, err)
-	require.Equal(t, 0, t0.NumActiveInvites())
-}
-
-func TestTeamCreateSeitanAndCancel(t *testing.T) {
-	testTeamCreateSeitanAndCancel(t, true /* seitanV1 */)
-	testTeamCreateSeitanAndCancel(t, false /* seitanV1 */)
 }
